+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # v.11.0
 
-* Added "tabindex", "order", "background-blend-mode", and "mix-blend-mode" properties 
-* Added GetTabIndex, GetOrder, GetBackgroundBlendMode, and GetMixBlendMode functions
+* Added "tabindex", "order", "column-fill", "background-blend-mode", and "mix-blend-mode" properties 
+* Added GetTabIndex, GetOrder, GetColumnFill, GetBackgroundBlendMode, and GetMixBlendMode functions
 * ClientItem, SetClientItem, and RemoveAllClientItems method added to Session interface
 * PropertyWithTag method of DataObject interface renamed to PropertyByTag
 
 # v.10.0
 
 * The Canvas.TextWidth method replaced by Canvas.TextMetrics
+* Added support of WebAssembly
 
 # v0.9.0